Push decode_raw down to simplates._decode
It belongs down there because it's not used elsewhere, it only makes
sense in the context of simplates.
